requirements.txt
certifi==2020.6.20
cycler==0.10.0
future==0.18.2
kiwisolver==1.2.0
matplotlib==3.3.2
mss==6.0.0
numpy==1.19.2
opencv-python==4.4.0.44
Pillow==8.0.1
pyparsing==2.4.7
PySide2==5.15.1
python-dateutil==2.8.1
PyYAML==5.3.1
shiboken2==5.15.1
six==1.15.0
torch==1.6.0
torchvision==0.7.0